+++ src/Support/ExecutionMiddleware/GraphqlExecutionMiddleware.php
@@ -7,6 +7,7 @@ use Closure;
 use GraphQL\Executor\ExecutionResult;
 use GraphQL\GraphQL as GraphQLBase;
 use GraphQL\Type\Schema;
+use Illuminate\Contracts\Config\Repository;
 use Rebing\GraphQL\Support\OperationParams;
 
 /**
@@ -14,11 +15,19 @@ use Rebing\GraphQL\Support\OperationParams;
  */
 class GraphqlExecutionMiddleware extends AbstractExecutionMiddleware
 {
+    /** @var Repository */
+    private $config;
+
+    public function __construct(Repository $config)
+    {
+        $this->config = $config;
+    }
+
     public function handle(string $schemaName, Schema $schema, OperationParams $params, $rootValue, $contextValue, Closure $next): ExecutionResult
     {
         $query = $params->getParsedQuery();
 
-        $defaultFieldResolver = config('graphql.defaultFieldResolver');
+        $defaultFieldResolver = $this->config->get('graphql.defaultFieldResolver');
 
         return GraphQLBase::executeQuery($schema, $query, $rootValue, $contextValue, $params->variables, $params->operation, $defaultFieldResolver);
     }